refactor(numbers): move HeaderValidation from Core to Numbers.SinchEvents namespace
